Streamlined the mempool crate and improved readability. It builds and all tests pass.
